refactorizing & fixes
