+++ src/components/coaching/Header.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import { Helmet } from 'react-helmet';
 import HeaderSelector from '../HeaderSelector';
+import LayoutButton from '../header/LayoutButton';
 
 class Header extends Component {
   state = {
@@ -40,13 +41,15 @@ class Header extends Component {
           >
             <span className="ec-header__logo">Ela Crain</span>
           </div>
-          <div className="ec-header__icons-wrapper">
-            {showLayoutIcons && (
-              <LayoutButton
-                onViewModeChange={onViewModeChange}
-                viewMode={viewMode}
-              />
-            )}
+          <div
+            className="ec-layout-button__wrapper"
+            style={{ visibility: showLayoutIcons ? 'visible' : 'hidden' }}
+          >
+            <LayoutButton
+              showLayoutIcons={showLayoutIcons}
+              onViewModeChange={onViewModeChange}
+              viewMode={viewMode}
+            />
           </div>
         </div>
       </header>
@@ -55,31 +58,3 @@ class Header extends Component {
 }
 
 export default Header;
-
-const LayoutButton = props => {
-  const { viewMode, onViewModeChange } = props;
-  const changeViewMode = () => {
-    viewMode === 'single'
-      ? onViewModeChange('overview')
-      : onViewModeChange('single');
-  };
-  return (
-    <div className="ec-layout-button" onClick={changeViewMode}>
-      {viewMode === 'single' ? (
-        <div className="ec-layout-button__content">
-          <div
-            className={`ec-layout-button__icon ec-layout-button__icon--overview`}
-          />
-          <span>List View</span>
-        </div>
-      ) : (
-        <div className="ec-layout-button__content">
-          <div
-            className={'ec-layout-button__icon ec-layout-button__icon--single'}
-          />
-          <span>Single View</span>
-        </div>
-      )}
-    </div>
-  );
-};
