+++ src/styles/_header.scss
@@ -6,20 +6,23 @@
 
     &--open {
         margin-top: 0;
+        transition: margin-top $animation-duration $animation-type;
     }
 
     &__wrapper {
         height: 80px;
-        width: 100%;
+
+        // width: 90%;
+        padding-left: 4%;
+        padding-right: 4%;
         display: flex;
-        justify-content: space-between;
-        flex-shrink: 0;
-        padding-left: 72px;
-        padding-right: 72px;
         align-items: center;
+        justify-content: space-between;
+   
     }
 
     &__links {
+        background: yellow;
         justify-content: flex-end;
         text-align: right;
         font-size: 18px;
@@ -27,6 +30,7 @@
         font-weight: normal;
         display: flex;
         align-items: center;
+        width: 33%;
 
         a {
             margin-left: 36px;
@@ -44,7 +48,9 @@
     }
 
     &__logo-container {
-        flex: 1;
+        background: green;
+        width: 154px;
+        // flex: auto;
         text-align: center;
         font-size: 50px;
         cursor: pointer;
@@ -63,6 +69,14 @@
 }
 
 .ec-layout-button {
+    &__wrapper {
+        width: 33%;
+        display: flex;
+        flex-flow: row;
+        justify-content: flex-start;
+        background: hotpink;
+    }
+
     width: 145px;
     height: 44px;
     border-radius: 22.5px;
@@ -133,7 +147,7 @@
         }
 
         &__links {
-            justify-content: flex-end;
+            justify-content: flex-start;
             text-align: right;
             font-size: 18px;
             letter-spacing: 1px;
@@ -158,6 +172,10 @@
     }
 
     .ec-layout-button {
+        &__wrapper {
+            justify-content: flex-end;
+        }
+
         border-color: $black;
 
         &__content {
